Add evaluator for Arrows (#173)
This evaluator is modelled as a transition function from 'input -> state -> (output, state)'.
It should be capable of handling all loop/delay/etc elements. The
semantics are currently that looping elements are evaluated once- i.e.
prevention of combinational loops is intended to be handled by external
lemmas. In the future this evaluator could be extended to handle combinational
loops that can be proved to converge after 'n' loops within the same cycle, if
this would be a useful property.