SubBytes - Vector.t -> list
In order to simplify the proof that sub_bytes and shift_rows commute,
it's better to have these two functions share the same representation.

Note that sub bytes is special in that the definition works whether the
data input is in row-major or column-major order. So, in the test here,
we interpret it to be a list of rows, and this makes its commutivity
with shift_rows definable without format conversions. Once we have mix
columns defined, that commutavity property might be defined without
format conversions as well, and we can prove 'transpose (sub_bytes x)
= sub_bytes (transpose x)'.