CavaClass: factor out indexConst (#711)
CavaClass represents the abstraction layer for some subset of our
circuits. Limiting this surface limits the ability of an implementation
of CavaClass to conflict with itself, and increases the usefulness of
proofs against a specific implementation of CavaClass (usually
combinational semantics).

For instance, one undocumented behavior of indexConst is that when the
index exceeds the bounds of the provided vector, the default signal is
returned. In the existing implementation, we have no guarantee that this
semantic is preserved between NetlistGeneration and our cominational
semantics.

In this case, unpackV is provides a strict superset of functionality to
indexConst. Specifically, lifting the use of an underlying index
function above the CavaClass function, based on the semantics of
unpackV, should produce the exact same underlying code.

Contrast this with other cases (such as and/xor/etc..) for which
replacing them with compositions of each other may produce more complex
circuits. For instance, imagine one FPGA uses a different basis of gates
to achieve functional completeness.

The hope is that this CL establishes a place and methodology to simplify
CavaClass. For maximum assurance far into the future, CavaClass would
also need to expose a list of properties which are required of the
underlying implementation, and we would want to base proofs off of this
instead of the combinational semantics (or we would want proofs that
other semantics are in some sense isomorphic to the combinational
cases). In either case, this simplification process reduces the API
surface we have to live/deal with.