Performance tweaks for Haskell execution (#728)
A bit of an extraction experiment: Having `nats` in extraction code is slow. This includes as constants and indices to vectors. I only modified SubBytes here as it was the easiest to modify, but it knocked 20s off `ghc --make`. Test vectors for `MixColumns` and `AddRoundKeyNetlist` are unmodified (they include vecs with nat indices) and still extract slow. Parallel ghc (`ghc -j`) halves the execution time.