define execution of risc-v machine with cava device as
interp of free monad (instead of directly using a Riscv monad
instance), so that we can prove the correspondence between
the compiler's risc-v machine and the cava risc-v machine for
each Riscv primitive separately, without having to go through
Run.run1 of all risc-v instructions.
However, this requires mid-conditions implying existence of
a state machine execution, and it's not clear how to make that
condition inductive.
Will probably have to require read_step/write_step in
mmioLoad/mmioStore of MinimalMMIO, instead of relying on post
asserting existence of an execution.