Add spec and proof of Hmac Inner block
The precondition enforces:
- SHA256 mode
- cmd_* signals don't change (where required)
- A non empty message to hash (required by the SHA circuit and proof)

I originally tried to use an updated Invariant.v class set where update_repr
was "weakend" to an arbitrary prop, but it was taking me to long.

The problem with the current equality/computable update_repr is that you can't
hide any of the inner circuit representation. e.g. SHA256 explicitly
defines SHA256_inner and padder representation and how it updates.  That is-
all circuit specifications require their repr to include a repr for all subcircuits,
and a computable update function for it. Whereas I believe this could be avoided with an
update_repr that could just posit some existential repr for subcircuits.

E.g. it shouldn't be required to track the explicitly SHA256 internal
state in the HMAC spec (even though the proof of it all follows from the
SHA256 spec), instead just that there exists some existential
representation which implies the subcircuit postcondition.

I found some difficulty in doing this though as alluded to.